added test class for LogService changed method getLogEntries(Class, int) to getLogEntries(Class, long) changed toString Method in class LogEntry

+++ src/main/java/de/randi2/utility/logging/LogService.java
@@ -15,5 +15,5 @@ public interface LogService {
 	
 	public List<String> getLogEntries(String username);
 	
-	public List<String> getLogEntries(Class<? extends AbstractDomainObject> clazz, int id);
+	public List<String> getLogEntries(Class<? extends AbstractDomainObject> clazz, long id);
 }
